+++ RELEASE-NOTES.md
@@ -1,5 +1,6 @@
 ## 4.0-M4
 
 * #1 OpenTelemetry bootstrap
-* #3 Injectable OTel Resource with Bootique application name #3
+* #2 Configurable SdkTracerProvider
+* #3 Injectable OTel Resource with Bootique application name